sam-api-post/requirements.txt
aws-psycopg2
requests
pandas
boto3
pymysql
numpy